Allow extended characters when parsing text
Extended characters are characters with an ASCII code above 127, like the latin letter E with acute. As the char is signed, this means that these characters have a negative value (which would mean a value that is below the space ' ' character).
This makes the comparison use an unsigned char instead to correctly handle extended characters.